+++ src/_logger.jl
@@ -6,23 +6,24 @@
 # Edit src/sysstate.yaml instead
 
 """
-    mutable struct Logger{P, O, D, L, W, T, S, F, Q}
+    mutable struct Logger{P, O, D, L, W, T, S, N, F, Q}
 
 Struct to store a simulation log. P is number of points of the tether, segments+1,
 O is the number of oriented frames, D is the number of twist surfaces, L is the
 number of pulleys, W is the number of winches, T is the number of tethers, S is
-the number of segments, F is the float type of the logged
-columns and Q is the number of time steps that will be pre-allocated.
+the number of segments, N is the number of aerodynamic panels, F is the float
+type of the logged columns and Q is the number of time steps that will be
+pre-allocated.
 
 Constructor:
 - Logger(P, steps; orients, deflections, pulleys, winches, tethers, segments,
-  precision)
+  panels, precision)
 
 Fields:
 
 $(TYPEDFIELDS)
 """
-@with_kw mutable struct Logger{P, O, D, L, W, T, S, F, Q}
+@with_kw mutable struct Logger{P, O, D, L, W, T, S, N, F, Q}
     points::Int64 = P
     orients::Int64 = O
     deflections::Int64 = D
@@ -30,6 +31,7 @@ $(TYPEDFIELDS)
     winches::Int64 = W
     tethers::Int64 = T
     segments::Int64 = S
+    panels::Int64 = N
     index::Int64 = 1
     time_vec::Vector{Float64} = zeros(Float64, Q)
     t_sim_vec::Vector{Float64} = zeros(Float64, Q)
@@ -67,8 +69,12 @@ $(TYPEDFIELDS)
     alpha4_vec::Vector{F} = zeros(F, Q)
     CL2_vec::Vector{F} = zeros(F, Q)
     CD2_vec::Vector{F} = zeros(F, Q)
-    aero_force_KA_vec::Vector{MVector{3, F}} = [zero(MVector{3, F}) for _ in 1:Q]
-    aero_moment_KA_vec::Vector{MVector{3, F}} = [zero(MVector{3, F}) for _ in 1:Q]
+    aero_force_KA_x_vec::Vector{MVector{O, F}} = [zero(MVector{O, F}) for _ in 1:Q]
+    aero_force_KA_y_vec::Vector{MVector{O, F}} = [zero(MVector{O, F}) for _ in 1:Q]
+    aero_force_KA_z_vec::Vector{MVector{O, F}} = [zero(MVector{O, F}) for _ in 1:Q]
+    aero_moment_KA_x_vec::Vector{MVector{O, F}} = [zero(MVector{O, F}) for _ in 1:Q]
+    aero_moment_KA_y_vec::Vector{MVector{O, F}} = [zero(MVector{O, F}) for _ in 1:Q]
+    aero_moment_KA_z_vec::Vector{MVector{O, F}} = [zero(MVector{O, F}) for _ in 1:Q]
     twist_angles_vec::Vector{MVector{D, F}} = [zero(MVector{D, F}) for _ in 1:Q]
     vel_kite_vec::Vector{MVector{3, F}} = [zero(MVector{3, F}) for _ in 1:Q]
     acc_vec::Vector{F} = zeros(F, Q)
@@ -86,6 +92,7 @@ $(TYPEDFIELDS)
     drag_force_y_vec::Vector{MVector{P, F}} = [zero(MVector{P, F}) for _ in 1:Q]
     drag_force_z_vec::Vector{MVector{P, F}} = [zero(MVector{P, F}) for _ in 1:Q]
     spring_force_vec::Vector{MVector{S, F}} = [zero(MVector{S, F}) for _ in 1:Q]
+    gamma_distribution_vec::Vector{MVector{N, F}} = [zero(MVector{N, F}) for _ in 1:Q]
     turn_rate_x_vec::Vector{MVector{O, F}} = [zero(MVector{O, F}) for _ in 1:Q]
     turn_rate_y_vec::Vector{MVector{O, F}} = [zero(MVector{O, F}) for _ in 1:Q]
     turn_rate_z_vec::Vector{MVector{O, F}} = [zero(MVector{O, F}) for _ in 1:Q]
@@ -95,6 +102,9 @@ $(TYPEDFIELDS)
     set_torque_vec::Vector{MVector{W, F}} = [zero(MVector{W, F}) for _ in 1:Q]
     set_speed_vec::Vector{MVector{W, F}} = [zero(MVector{W, F}) for _ in 1:Q]
     set_force_vec::Vector{MVector{W, F}} = [zero(MVector{W, F}) for _ in 1:Q]
+    set_ext_force_x_vec::Vector{MVector{P, F}} = [zero(MVector{P, F}) for _ in 1:Q]
+    set_ext_force_y_vec::Vector{MVector{P, F}} = [zero(MVector{P, F}) for _ in 1:Q]
+    set_ext_force_z_vec::Vector{MVector{P, F}} = [zero(MVector{P, F}) for _ in 1:Q]
     var_01_vec::Vector{F} = zeros(F, Q)
     var_02_vec::Vector{F} = zeros(F, Q)
     var_03_vec::Vector{F} = zeros(F, Q)